# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Social network platform API using bounded context architecture",
        "title": "DarkVoid API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@darkvoid.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/hashtags/search": {
            "get": {
                "description": "Returns up to 10 hashtag names that start with the given prefix (min 2 chars). Results cached 2 minutes.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "hashtags"
                ],
                "summary": "Search hashtags by prefix",
                "operationId": "searchHashtags",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Prefix to search (min 2 chars, without #)",
                        "name": "q",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.HashtagSearchResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/hashtags/trending": {
            "get": {
                "description": "Returns the top trending hashtags based on post usage in the last 24 hours",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "hashtags"
                ],
                "summary": "Trending hashtags",
                "operationId": "getTrendingHashtags",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.TrendingHashtagListResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/hashtags/{name}/posts": {
            "get": {
                "description": "Returns cursor-paginated public posts for a hashtag",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "hashtags"
                ],
                "summary": "Posts by hashtag",
                "operationId": "getPostsByHashtag",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Hashtag name (without #)",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Pagination cursor",
                        "name": "cursor",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Max results (default 20)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.HashtagPostListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/posts/{postID}/like": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Like the post if not yet liked, unlike if already liked. Returns the new like state.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "posts"
                ],
                "summary": "Toggle like on a post",
                "operationId": "toggleLike",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Post ID (UUID)",
                        "name": "postID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "liked: true if now liked, false if now unliked",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid post ID or self-like",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.AuthorResponse": {
            "type": "object",
            "properties": {
                "avatar_url": {
                    "type": "string"
                },
                "display_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "dto.HashtagPostListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.PostResponse"
                    }
                },
                "hashtag": {
                    "type": "string"
                },
                "next_cursor": {
                    "type": "string"
                }
            }
        },
        "dto.HashtagSearchResponse": {
            "type": "object",
            "properties": {
                "prefix": {
                    "type": "string"
                },
                "results": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "dto.MediaResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "media_type": {
                    "type": "string"
                },
                "position": {
                    "type": "integer"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "dto.MentionResponse": {
            "type": "object",
            "properties": {
                "display_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "dto.PostResponse": {
            "type": "object",
            "properties": {
                "author": {
                    "$ref": "#/definitions/dto.AuthorResponse"
                },
                "author_id": {
                    "type": "string"
                },
                "comment_count": {
                    "type": "integer"
                },
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_following_author": {
                    "type": "boolean"
                },
                "is_liked": {
                    "type": "boolean"
                },
                "like_count": {
                    "type": "integer"
                },
                "media": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.MediaResponse"
                    }
                },
                "mentions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.MentionResponse"
                    }
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "updated_at": {
                    "type": "string"
                },
                "visibility": {
                    "type": "string"
                }
            }
        },
...
        "dto.TrendingHashtagListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.TrendingHashtagResponse"
                    }
                }
            }
        },
        "dto.TrendingHashtagResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "errors.ErrorDetail": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "errors.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/errors.ErrorDetail"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}